feat: dialog is now its own component
